* Fix refactored event creation.
